# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Golang boilerplate API with common features",
        "title": "Go Boilerplate API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/payments/callback": {
            "post": {
                "description": "Receives HTTP POST notification from Midtrans when transaction status changes. This URL must be registered in Midtrans Dashboard \u003e Settings \u003e Payment Notification URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Midtrans payment notification webhook",
                "parameters": [
                    {
                        "description": "Midtrans notification payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payments/create": {
            "post": {
                "description": "Receives order data from WhatsApp Bot, generates a Midtrans Snap transaction, saves to DB, and returns a payment URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_service_payment.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-boilerplate_internal_service_payment.CreatePaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/payments/process": {
            "post": {
                "description": "Receives snap.js callback result from frontend, verifies with Midtrans API, and updates transaction status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment result from frontend",
                "parameters": [
                    {
                        "description": "Payment result from snap.js",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_service_payment.PaymentResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-boilerplate_internal_service_payment.PaymentStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/payments/status/{order_id}": {
            "get": {
                "description": "Checks real-time payment status from Midtrans API with database fallback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Check payment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-boilerplate_internal_service_payment.PaymentStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/payments/wa-flow-endpoint": {
            "post": {
                "description": "Handles data_exchange action from WhatsApp Flow, forwards order form data to SUMMARY_ORDER screen",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Flow"
                ],
                "summary": "WhatsApp Flow data exchange endpoint",
                "parameters": [
                    {
                        "description": "WhatsApp Flow request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_payment.WAFlowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_payment.WAFlowResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/xample": {
            "get": {
                "description": "Returns example data for testing purposes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Example"
                ],
                "summary": "Get example data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-boilerplate_internal_common_type.ResponseAPI": {
            "type": "object",
            "properties": {
                "data": {},
                "debug": {
                    "$ref": "#/definitions/go-boilerplate_internal_common_type.ResponseAPIDebug"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_common_type.ResponseAPIDebug": {
            "type": "object",
            "properties": {
                "endTime": {
                    "description": "ISO8601 format for consistency with StartTime",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "runtimeMs": {
                    "description": "Runtime in milliseconds for better precision",
                    "type": "integer"
                },
                "startTime": {
                    "description": "ISO8601 format, e.g., \"2025-01-09T15:04:05Z07:00\"",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_service_payment.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "customer": {
                    "$ref": "#/definitions/go-boilerplate_internal_service_payment.CustomerInfo"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/go-boilerplate_internal_service_payment.ItemDetail"
                    }
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "order_id": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_service_payment.CreatePaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_url": {
                    "type": "string"
                },
                "snap_token": {
                    "type": "string"
                },
                "snap_url": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_service_payment.CustomerInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_service_payment.ItemDetail": {
            "type": "object",
            "required": [
                "id",
                "name",
                "price",
                "qty"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
        "go-boilerplate_internal_service_payment.PaymentResultRequest": {
            "type": "object",
            "properties": {
                "gross_amount": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "signature_key": {
                    "type": "string"
                },
                "status_code": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_status": {
                    "type": "string"
                }
            }
        },
        "go-boilerplate_internal_service_payment.PaymentStatusResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "internal_handler_payment.WAFlowOrderData": {
            "type": "object",
            "properties": {
                "alamat_lengkap": {
                    "type": "string"
                },
                "items_text": {
                    "type": "string"
                },
                "kode_pos": {
                    "type": "string"
                },
                "kota_kecamatan": {
                    "type": "string"
                },
                "nama_penerima": {
                    "type": "string"
                },
                "nomor_handphone": {
                    "type": "string"
                },
                "provinsi": {
                    "type": "string"
                },
                "total_barang": {
                    "type": "string"
                },
                "total_biaya": {
                    "type": "string"
                },
                "total_pengiriman": {
                    "type": "string"
                }
            }
        },
        "internal_handler_payment.WAFlowRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/internal_handler_payment.WAFlowOrderData"
                }
            }
        },
        "internal_handler_payment.WAFlowResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_handler_payment.WAFlowOrderData"
                },
                "screen": {
                    "type": "string"
                }
            }
        }
    }
}